Add Elixir.tmbundle
I have never created a Homebrew Cask before so I'm not sure if this is
correct or not. I used some examples in this repo and I also used the
"Adding a Cask" documentation.

Please let me know what I should change or how I should test this to
make sure it works properly.